IndexSelectorを別ファイルに分離
